Implement Gateway v4, refactor to accommodate two protocols
